Move end-to-end test to module with other tip test
